requirements.txt
django
djangorestframework
django-cors-headers
drf-yasg
gunicorn
psycopg2
django-cloudinary-storage
python-decouple
django-cors-headers
mysqlclient
drf-base64
pillow
django-psdb-engine
djangorestframework-simplejwt